fix(regroup): unify BySizeRoundedPerComponent on the v3 geometry home
The merge of origin/main brought main's standalone cfdmod.regroup module
alongside the v3 cfdmod.geometry.grouping.regroup spec, leaving two
distinct classes named BySizeRoundedPerComponent. The top-level package
re-exported the geometry one while RegroupConfig's union referenced the
regroup-local one, so RegroupConfig rejected the exported spec (6 remesh
fixture tests errored with a pydantic model_type mismatch).

Make cfdmod.geometry.grouping.regroup the single source of truth per
issue #131's de-duplication intent:
- promote the geometry spec to the superset: add the restrict_to field
  and the "at least one target axis" validator from main's version, and
  honor restrict_to in expand_size_rounded_chain (no-op when None, so the
  v3 regroup_topology op behavior is unchanged).
- cfdmod.regroup.parameters now imports BySizeRoundedPerComponent and
  RegroupSpec from the geometry home instead of redefining them.

The two chain expanders (expand_size_rounded_chain vs the regroup
module's transformation-aware expand_regroup_chain) stay separate: they
serve different consumers and only the spec class needed unifying.
